feat: Fake Server and API for character is clarified.
