+++ app/services/api/characters/Fake/data.ts
@@ -1,0 +1,26 @@
+import * as Types from '../api.types'
+
+export const DEFAULT_DATA: Types.Character[] = [
+  {
+    id: '11111-111111-11111-11111',
+    initiative: 0,
+    roll: 0,
+    dexterity: 12,
+    modifiers: 18,
+    name: 'Dirby',
+    avatar: {
+      thumbnail: 'https://randomuser.me/api/portraits/men/1.jpg'
+    }
+  },
+  {
+    id: '11111-111111-11111-11112',
+    initiative: 0,
+    roll: 0,
+    dexterity: 6,
+    modifiers: 2,
+    name: 'Yamanu',
+    avatar: {
+      thumbnail: 'https://randomuser.me/api/portraits/women/80.jpg'
+    }
+  }
+]
